cuol-wordle-backend/cuol_wordle_django/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.4
PyJWT==1.7.1
sqlparse==0.4.3
